refactor: restructure the components around the project
refactor: rename confussing component

refactor: restructure the components around the project

refactor: rename confussing component

refactor: refactor issue
