+++ src/components/Configs/Insights/ConfigInsightsColumns.tsx
@@ -1,7 +1,7 @@
+import { ConfigAnalysis, ConfigItem } from "@flanksource-ui/api/types/configs";
+import { DateCell } from "@flanksource-ui/ui/DataTable/Cells/DateCells";
 import { ColumnDef } from "@tanstack/react-table";
 import { Link } from "react-router-dom";
-import { ConfigAnalysis, ConfigItem } from "../../../api/types/configs";
-import { DateCell } from "../../../ui/table";
 import { ConfigIcon } from "../../Icon/ConfigIcon";
 import ConfigInsightsIcon from "./ConfigInsightsIcon";
 import ConfigInsightsSeverityIcons from "./ConfigInsightsSeverityIcons";
