+++ src/components/JobsHistory/JobsHistoryTableColumn.tsx
@@ -1,10 +1,10 @@
+import { DateCell } from "@flanksource-ui/ui/DataTable/Cells/DateCells";
 import { ColumnDef } from "@tanstack/react-table";
 import dayjs from "dayjs";
 import duration from "dayjs/plugin/duration";
 import relativeTime from "dayjs/plugin/relativeTime";
 import { GoCopy, GoLinkExternal } from "react-icons/go";
 import { Link, useSearchParams } from "react-router-dom";
-import { DateCell } from "../../ui/table";
 import { formatJobName } from "../../utils/common";
 import { formatDuration } from "../../utils/date";
 import { useCopyToClipboard } from "../Hooks/useCopyToClipboard";
