+++ src/components/LogBackends/LogBackendsList.tsx
@@ -1,7 +1,7 @@
 import { Avatar } from "@flanksource-ui/ui/Avatar";
 import { DataTable } from "@flanksource-ui/ui/DataTable";
+import { DateCell } from "@flanksource-ui/ui/DataTable/Cells/DateCells";
 import { Modal } from "@flanksource-ui/ui/Modal";
-import { DateCell } from "@flanksource-ui/ui/table";
 import { CellContext, ColumnDef } from "@tanstack/react-table";
 import { useState } from "react";
 import { LogBackends } from "./LogBackends";
