--- src/ui/table/index.tsx
@@ -1,18 +1,0 @@
-import {
-  DateCell,
-  CreatedAtCell,
-  UpdatedAtCell,
-  DeletedAtCell
-} from "./DateCells";
-import { TagsCell } from "./TagCell";
-import { UserCell, AvatarCell } from "./User";
-
-export {
-  DateCell,
-  TagsCell,
-  CreatedAtCell,
-  DeletedAtCell,
-  UpdatedAtCell,
-  UserCell,
-  AvatarCell
-};
